move header of AlignmentStats into a single file implementation
